chore: clean up all Rust warnings across workspace
- Fix 101 hidden lifetime parameter warnings (add explicit `'_`)
- Fix ~129 clone-on-Copy warnings (Span, Symbol, Value, Visibility)
- Fix ~75 unnecessary cast warnings (u16→u16, usize→usize, raw ptrs)
- Fix ~39 redundant closure warnings (replace with fn references)
- Fix ~19 needless return warnings (remove explicit returns)
- Fix ~20 args.get(0) → args.first() warnings
- Fix ~15 map_or simplification warnings
- Fix 13 redundant match guard warnings
- Fix ~20 unused imports/variables/dead code items
- Fix 3 unreachable pattern warnings
- Fix deprecated Gc type alias usage
- Suppress deprecated libc::mach_host_self/mach_task_self with allow
- Fix various other clippy lints (io::Error::other, strip_prefix,
  unwrap_or_default, manual_flatten, transmute annotations, etc.)

Clean up crate-level allow attributes:
- Remove 18 blanket #[allow] with TODO comment from lib.rs
- Remove duplicate 18-line block from vm/mod.rs
- Add 4 justified crate-level allows with explanatory comments
- Add per-site #[allow] with justification for intentional patterns

Result: zero warnings on cargo check and cargo clippy (excluding
missing_docs). All 3,800+ tests pass across all crates and features.

117 files changed, 850 insertions, 980 deletions (net -130 lines)

+++ crates/raya-stdlib-posix/src/os.rs
@@ -61,7 +61,7 @@ pub fn free_memory(_ctx: &dyn NativeContext, _args: &[NativeValue]) -> NativeCal
 
 /// Get system uptime in seconds
 pub fn uptime(_ctx: &dyn NativeContext, _args: &[NativeValue]) -> NativeCallResult {
-    NativeCallResult::f64(get_uptime() as f64)
+    NativeCallResult::f64(get_uptime())
 }
 
 /// Get OS line ending
@@ -376,6 +376,7 @@ fn get_free_memory() -> u64 {
         let page_size = unsafe { libc::sysconf(libc::_SC_PAGESIZE) } as u64;
         let mut vm_stat: libc::vm_statistics64 = unsafe { mem::zeroed() };
         let mut info_count = (mem::size_of::<libc::vm_statistics64>() / mem::size_of::<libc::natural_t>()) as libc::mach_msg_type_number_t;
+        #[allow(deprecated)] // libc deprecates in favor of mach2 crate, but we use libc
         let ret = unsafe {
             libc::host_statistics64(
                 libc::mach_host_self(),
